Updated to 0.1.5 (to the development branch) (#28)
<!--
For Work In Progress Pull Requests, please use the Draft PR feature,
see https://github.blog/2019-02-14-introducing-draft-pull-requests/ for
further details.

     For a timely review/response, please avoid force-pushing additional
     commits if your PR already received reviews or comments.

Before submitting a Pull Request, please ensure you've done the
following:
- 📖 Read the our Code of Conduct:
https://github.com/ekinkaradag/snake-vue3/blob/main/.github/CODE_OF_CONDUCT.md
     - 👷‍♀️ Create small PRs. In most cases this will be possible.
     - ✅ Provide tests for your changes (Not applicable yet. Disregard).
     - 📝 Use descriptive commit messages.
- 📗 Update any related documentation and include any relevant
screenshots.
-->

## What type of PR is this? (check all applicable)

- [X] 💻 Refactor
- [X] 🍕 Feature
- [X] 🐛 Bug Fix
- [ ] 🚀 Optimization
- [ ] 📃 Documentation Update

## Description

## Related Tickets & Documents

<!--
For pull requests that relate or close an issue, please include them
below. We like to follow [Github's guidance on linking issues to pull
requests](https://docs.github.com/en/issues/tracking-your-work-with-issues/linking-a-pull-request-to-an-issue).

For example having the text: "closes #1234" would connect the current
pull
request to issue 1234.  And when we merge the pull request, Github will
automatically close the issue.
-->

## QA Instructions, Screenshots, Recordings

_Please replace this line with instructions on how to test your changes,
a note
on the devices and browsers this has been tested on, as well as any
relevant
images for UI changes._

### UI accessibility checklist

_If your PR includes UI changes, please utilize this checklist:_

- [X] Semantic HTML implemented?
- [X] Keyboard operability supported?
- [X] Color contrast tested?

## [disregard] Added/updated tests? (Not applicable yet)

- [ ] 👍 Yes
- [ ] 👎 No, and this is why: _please replace this line with details on
why tests
      have not been included_
- [ ] ✋ I need help with writing tests

## [optional] Are there any post deployment tasks we need to perform?

## [optional] What gif best describes this PR or how it makes you feel?